Move enums to module pretix/common
